travel certicates 2017/2018

+++ hurumap_ng/sql/travel_certificates.sql
@@ -28,7 +28,7 @@ CREATE TABLE public.travel_certificates (
     geo_code character varying(10) NOT NULL,
     geo_version character varying(128) NOT NULL,
     year character varying(128) NOT NULL,
-    total numeric
+    total integer
 );
